pyproject.toml
[project]
name = "node-red"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "hf-xet>=1.0.3",
    "kokoro>=0.7.9",
    "llama-cpp-python>=0.3.8",
    "loguru>=0.7.3",
    "pydantic>=2.11.3",
    "pypdf2[crypto]>=3.0.1",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
    "torch>=2.6.0",
    "numpy>=1.22",
    "pip>=25.0.1",
    "huggingface-hub>=0.30.2",
    "soundfile>=0.13.1",
    "tts>=0.22.0",
    "python-docx>=1.1.2",
    "argparse>=1.4.0",
    "bs4>=0.0.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
